Custom borders

+++ Font.cs
@@ -5,8 +5,8 @@ namespace Tephanik
         public string? fontkey { get; set; }
         public string? type { get; set; }
         public string? name { get; set; }
-        public int up { get; set; }
-        public int ut { get; set; }
+        public double up { get; set; }
+        public double ut { get; set; }
         public List<(char, int)> cw { get; set; }
         public string enc { get; set; }
         public List<(int, dynamic)> uv { get; set; }
